test : use standard types

+++ mytest.c
@@ -14,8 +14,9 @@
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include<stdio.h>
-#include"mythreads.h"
+#include <stdio.h>
+#include <inttypes.h>
+#include "mythreads.h"
 
 void longJob(void *dummy){
 	volatile unsigned long long i;
@@ -23,14 +24,14 @@ void longJob(void *dummy){
 }
 
 int main(){
-	int size;
+	uint64_t size;
 	printf("\nEnter the number of threads to start from : ");
-	scanf("%d", &size);
+	scanf("%" SCNu64 , &size);
 	ThreadPool *pool = createPool(size);
 	printf("\nEnter number of jobs : ");
-	scanf("%d", &size);
-	int i = 0;
-	for(i=0;i<size;i++)
+	scanf("%" SCNu64, &size);
+	uint64_t i = 0;
+	for(i = 0;i < size;i++)
 		addJobToPool(pool, &longJob, NULL);
 	char choice = '1';
 	while(choice>='0' && choice<'9'){
@@ -47,17 +48,17 @@ int main(){
 		scanf(" %c", &choice);
 		switch(choice){
 			case '0': printf("\n[CHOICE:INPUT] Number of jobs : ");
-				  scanf("%d", &size);
+				  scanf("%" SCNu64, &size);
 				  for(i=0;i<size;i++)
 					  addJobToPool(pool, &longJob, NULL);
 				  break;
-			case '1': printf("\n[CHOICE:INFO] Pending jobs %lu", getJobCount(pool));
+			case '1': printf("\n[CHOICE:INFO] Pending jobs %" PRIu64 , getJobCount(pool));
 				  break;
 			case '2': addThreadsToPool(pool, 1);
 				  break;
 			case '3': removeThreadFromPool(pool);
 				  break;
-			case '4': printf("\n[CHOICE:INFO] Number of threads %u", getThreadCount(pool));
+			case '4': printf("\n[CHOICE:INFO] Number of threads %" PRIu64, getThreadCount(pool));
 				  break;
 			case '5': waitToComplete(pool);
 				  break;
